if farmer chooses regional language then if possible tru to give that weather advisor matter also in seleected regional language, and also the ai price suggestion also in selected language.
and in the farming news, only english words are reading out, makesure that is also reads selected regional language also. and there itself the stop reading buttion isnot working, once started reading it is not stoping inbetween, correct it also.

and also try to give the lastest news.. and also in farmers login in the analystics the ai sales and recommendation only giving in english, if possible try to give in farmer selected language